requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
audio-recorder-streamlit
requests>=2.28.0
sentence-transformers
langchain-huggingface
langchain-community